Revert "add arm64 support"
