+++ .gitignore
@@ -5,6 +5,3 @@ server/prometheus_data/
 .vscode/
 .build-version
 test/dummy/*
-.env
-grafana.db
-
