--- agent/cita_exporter/.dockerignore
@@ -1,1 +1,0 @@
-.git
