+++ agent/docker-compose.yml
@@ -2,7 +2,7 @@ version: "3"
 services:
 
   citamon_agent_host_exporter:
-    image: prom/node-exporter:v0.18.0
+    image: prom/node-exporter:v0.17.0
     hostname: host_exporter
     container_name: citamon_agent_host_exporter
     volumes:
@@ -15,19 +15,19 @@ services:
       - citamon-agent-net
 
   citamon_agent_process_exporter:
-    image: citamon/agent-process-exporter:0.4.0
+    image: ncabatoff/process-exporter:0.4.0
     hostname: process_exporter
     container_name: citamon_agent_process_exporter
     volumes:
       - /proc:/host/proc
-      - ./process_exporter/config:/config
+      - ./process_exporter_config:/config
       - /etc/localtime:/etc/localtime
     command: "--procfs /host/proc --config.path /config/process_list.yml"
     networks:
       - citamon-agent-net
 
   citamon_agent_rabbitmq_exporter:
-    image: citamon/agent-rabbitmq-exporter:1.0.0-RC
+    image: kbudde/rabbitmq-exporter:1.0.0-RC
     hostname: rabbitmq_exporter
     container_name: citamon_agent_rabbitmq_exporter
     network_mode: "host"
@@ -37,13 +37,15 @@ services:
       - "RABBIT_CAPABILITIES=bert,no_sort"
 
   citamon_agent_cita_exporter:
-    image: citamon/agent-cita-exporter:20.2.2
+    build: ./cita_exporter
+    image: citamon/agent-cita-exporter
     hostname: ${HOSTNAME}
     container_name: citamon_agent_cita_exporter_${NODE_IP}_${NODE_PORT}
     pid: "host"
     volumes:
       - ${NODE_DIR}:${NODE_DIR}:ro
       - ${SOFT_PATH}:${SOFT_PATH}:ro
+      - ./cita_monitor_agent.py:/config/cita_monitor_agent.py:ro
       - /etc/localtime:/etc/localtime
     environment:
       - NODE_IP_PORT=${NODE_IP}:${NODE_PORT}
