--- agent/process_exporter/config/process_list.yml
@@ -1,12 +1,0 @@
-process_names:
-  - comm:
-    - cita-forever
-    - cita-network
-    - cita-jsonrpc
-    - cita-bft
-    - cita-chain
-    - cita-auth
-    - cita-executor
-  - name: "{{.Comm}}"
-    cmdline:
-    - 'rabbitmq-server'
