+++ server/grafana/Dockerfile
@@ -1,4 +1,4 @@
-FROM grafana/grafana:5.4.3
+FROM grafana/grafana:5.4.2
 
 USER root
